model/requirements.txt
altair==4.2.0
aniso8601==9.0.1
attrs==22.1.0
bcrypt==4.0.1
blinker==1.5
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
commonmark==0.9.1
dataclasses==0.6
decorator==5.1.1
django-js-asset==1.2.2
entrypoints==0.4
extra-streamlit-components==0.1.56
Flask==2.2.3
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
gitdb==4.0.10
GitPython==3.1.29
greenlet==2.0.2
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.1
MarkupSafe==2.1.2
numpy==1.23.5
packaging==21.3
pandas==1.5.2
Pillow==9.3.0
protobuf==3.20.1
pyarrow==10.0.1
pydeck==0.8.0
Pygments==2.13.0
PyJWT==2.6.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.1
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.2
rich==12.6.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.9
sqlparse==0.2.4
streamlit==1.15.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.5.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.15
validators==0.20.0
Werkzeug==2.2.3
zipp==3.15.0
